Register `ty.runTest` as a command
